+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## next
 - Set minimal supported Gradle version to 6.7
 - Fix for `getHeaderField("Location") must not be null` [#960](../../issues/960)
+- Added `BuildFeature` feature flags
 
 ## 1.5.3
 - Updated dependencies marked as vulnerable [#459](../../issues/459)